Started on better plinko
Plinko system now runs a baked physics sim in python and saves to a
file. This allows for calculable and repeatable plinko sims. It also
removes a lot of the heavy lifting from the client, and if the rendered
to a gif/video, could even remove almost all the heavy lifting.
The API barebones are there but still needs things like who triggered
the plinko run.